# api-spec/src/main/resources/oas3.0.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /todos/{id}:
    get:
      tags:
      - to-do-controller
      operationId: retrieveItem
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ToDoItem'
    put:
      tags:
      - to-do-controller
      operationId: updateItem
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToDoItem'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - to-do-controller
      operationId: deleteItem
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /todos:
    get:
      tags:
      - to-do-controller
      operationId: retrieveAllItems
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToDoItem'
    post:
      tags:
      - to-do-controller
      operationId: createItem
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToDoItem'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
components:
  schemas:
    ToDoItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        completed:
          type: boolean